# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Que hay pa' hacer API
  version: 0.2.0
  description: API contract for MVP mobile (Bogota)
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/events:
    get:
      tags: [events]
      summary: Discovery feed with ranking and filters
      operationId: listDiscoveryEvents
      parameters:
        - in: query
          name: categories
          schema:
            type: string
        - in: query
          name: cityId
          schema:
            type: string
        - in: query
          name: dateFrom
          schema:
            type: string
            format: date-time
        - in: query
          name: dateTo
          schema:
            type: string
            format: date-time
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: radiusKm
          schema:
            type: number
      responses:
        '200':
          description: Ranked feed
...
  /api/v1/events/{id}:
    get:
      tags: [events]
      summary: Event detail
      operationId: getEventById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event detail
  /api/v1/events/{id}/bookmark:
    post:
      tags: [bookmarks]
      summary: Bookmark event
      operationId: bookmarkEvent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Event bookmarked
    delete:
      tags: [bookmarks]
      summary: Remove bookmark
      operationId: unbookmarkEvent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event unbookmarked
  /api/v1/events/{id}/report:
    post:
      tags: [reports]
      summary: Report event
      operationId: reportEvent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Event reported
...
  /api/v1/admin/reports/{id}:
    patch:
      tags: [moderation]
      summary: Update report status
      operationId: updateReportStatus
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report updated